--- src/main/java/com/srans/nestserver/repository/SransUserRepository.java
@@ -1,16 +1,0 @@
-package com.srans.nestserver.repository;
-
-
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import com.srans.nestserver.model.SransUser;
-
-@Repository
-public interface SransUserRepository extends JpaRepository<SransUser, Long>{
-
-	
-
-
-}
