+++ src/main/java/com/srans/nestserver/repository/TenantRepository.java
@@ -1,0 +1,16 @@
+package com.srans.nestserver.repository;
+
+
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import com.srans.nestserver.model.Tenant;
+
+@Repository
+public interface TenantRepository extends JpaRepository<Tenant, Long>{
+
+	
+
+
+}
