+++ src/main/resources/application.properties
@@ -1,7 +1,7 @@
 
 spring.datasource.url=jdbc:postgresql://localhost:5432/nest?createDatabaseIfNotExist=true
 spring.datasource.username=postgres
-spring.datasource.password=admin
+spring.datasource.password=Hello@123
 spring.jpa.show-sql = true
